# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/projects/{project_id}/runs": {
      "post": {
        "operationId": "createSimulationRun",
        "parameters": [
          {
            "name": "project_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRunCreateDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor is not authorized.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource is not visible.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "The command conflicts with durable resource state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "413": {
            "description": "The request exceeds the bounded body envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "415": {
            "description": "A command did not use application/json.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "The request is invalid or outside the supported scope.\n\nThe request is invalid or outside the supported scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "A durable rate limit was reached.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v2/projects/{project_id}/behavioral-demo-runs": {
      "post": {
        "description": "Runs the fixed, visibly synthetic authored-demo audience. The output is experimental and is not a population estimate.",
        "operationId": "createBehavioralDemoRun",
        "parameters": [
          {
            "name": "project_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehavioralDemoRunCreateDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor is not authorized.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource is not visible.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "The command conflicts with durable resource state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "413": {
            "description": "The request exceeds the bounded body envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "415": {
            "description": "A command did not use application/json.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "The request is invalid or outside the supported scope.\n\nThe request is invalid or outside the supported scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "A durable rate limit was reached.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Create an experimental synthetic behavioral simulation",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v2/runs/{run_id}": {
      "get": {
        "operationId": "getSimulationRun",
        "parameters": [
          {
            "name": "run_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor is not authorized.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource is not visible.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "The command conflicts with durable resource state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "413": {
            "description": "The request exceeds the bounded body envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "415": {
            "description": "A command did not use application/json.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "The request is invalid or outside the supported scope.\n\nThe request is invalid or outside the supported scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "A durable rate limit was reached.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v2/runs/{run_id}/audit-history": {
      "get": {
        "description": "Returns bounded state transitions without actor identities, payloads, prompts, agent memory, rationale, or free-form metadata.",
        "operationId": "getRunAuditHistory",
        "parameters": [
          {
            "name": "run_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAuditHistoryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor is not authorized.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource is not visible.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "The command conflicts with durable resource state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "413": {
            "description": "The request exceeds the bounded body envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "415": {
            "description": "A command did not use application/json.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "The request is invalid or outside the supported scope.\n\nThe request is invalid or outside the supported scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "A durable rate limit was reached.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "A required dependency is temporarily unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Get the sanitized durable run state history",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIMULA Control Plane",
    "description": "Migration-safe SIMULA control-plane contract. Behavioral outputs remain experimental until independently validated.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProblemFieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code"
        ],
        "additionalProperties": false
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "correlation_id"
        ],
        "additionalProperties": false
      },
...
      "SimulationRunCreateDto": {
        "type": "object",
        "properties": {
          "stimulus_version_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "stimulus_version_id"
        ],
        "additionalProperties": false
      },
...
      "SimulationRunResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "stimulus_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "audience_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "retrying",
              "cancel_requested",
              "succeeded",
              "failed",
              "canceled"
            ]
          },
          "schema_version": {
            "type": "number",
            "enum": [
              1,
              2
            ]
          },
          "dispatch_generation": {
            "type": "number",
            "minimum": 1,
            "maximum": 3
          },
          "job_id": {
            "oneOf": [
              {
                "pattern": "^run:[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}:dispatch:[1-3]$"
              },
              {
                "pattern": "^run-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}-generation-[1-3]$"
              }
            ]
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "failure": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SimulationRunFailureDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "organization_id",
          "project_id",
          "stimulus_version_id",
          "audience_version_id",
          "state",
          "schema_version",
          "dispatch_generation",
          "job_id",
          "version",
          "created_at"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
}